Triggering wallect connection from trade page on submission
